fix(payload): Accept custom payloads
Adds `.custom` method to payload helpers, also loosening type checking, to allow any object to be
given and become the payload which goes through to the message adapter, without requiring it to fit
the internal attachment/action interfaces. The adapter still needs to recognise the payload
attributes, or it will not be able to convert them to the platform's required message object.

Fixes #100